requirements.txt
absl-py==0.7.1
actionlib==1.11.13
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.2
angles==1.9.11
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
astropy==3.1.2
atomicwrites==1.2.1
attrs==18.2.0
audioread==2.1.6
azure-cognitiveservices-speech==1.5.1
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bitarray==0.8.3
bkcharts==0.2
bleach==3.0.0
bokeh==1.0.4
bondpy==1.8.3
boto==2.49.0
Bottleneck==1.2.1
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
catkin==0.7.14
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
Click==7.0
clipboard==0.0.4
cloudpickle==0.6.1
clyent==1.2.2
colorama==0.4.1
contextlib2==0.5.5
cryptography==2.6.1
cv-bridge==1.12.8
cycler==0.10.0
Cython==0.29.6
cytoolz==0.9.0.1
dask==0.19.4
decorator==4.3.0
defusedxml==0.5.0
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
distributed==1.26.0
docutils==0.14
dynamic-reconfigure==1.5.50
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==1.0.2
gast==0.2.2
gazebo-plugins==2.5.17
gazebo-ros==2.5.17
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.7
gevent==1.4.0
gmpy2==2.0.8
greenlet==0.4.15
grpcio==1.21.1
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.8
image-geometry==1.12.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.0.0
interactive-markers==1.11.4
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.16
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.3
jeepney==0.4
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==18.0.0
kiwisolver==1.0.1
laser-geometry==1.6.4
lazy-object-proxy==1.3.1
llvmlite==0.28.0
locket==0.2.0
lxml==4.3.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mavros==0.27.0
mccabe==0.6.1
message-filters==1.12.14
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
mock==3.0.5
more-itertools==6.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==5.7.8
numba==0.43.1
numexpr==2.6.9
numpy==1.16.2
numpydoc==0.8.0
olefile==0.46
openpyxl==2.6.1
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
partd==0.3.10
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.9.0
pluginlib==1.11.3
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.6.1
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.1.1
Pygments==2.3.1
pylint==2.3.1
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyperclip==1.7.0
PyQt5==5.12
PyQt5-sip==4.19.17
pyrsistent==0.14.11
PySocks==1.6.8
PySoundFile==0.9.0.post1
pytest==4.3.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
python-qt-binding==0.3.4
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
pyzmq==18.0.0
qt-dotgraph==0.3.11
qt-gui==0.3.11
qt-gui-cpp==0.3.11
qt-gui-py-common==0.3.11
QtAwesome==0.5.7
qtconsole==4.4.3
QtPy==1.7.0
requests==2.21.0
resource-retriever==1.12.4
rope==0.12.0
ros-numpy==0.0.2
rosbag==1.12.14
rosboost-cfg==1.14.4
rosclean==1.14.4
roscreate==1.14.4
rosgraph==1.12.14
roslaunch==1.12.14
roslib==1.14.4
roslint==0.11.0
roslz4==1.12.14
rosmake==1.14.4
rosmaster==1.12.14
rosmsg==1.12.14
rosnode==1.12.14
rosparam==1.12.14
rospy==1.12.14
rosserial-client==0.7.7
rosserial-python==0.7.7
rosservice==1.12.14
rostest==1.12.14
rostopic==1.12.14
rosunit==1.14.4
roswtf==1.12.14
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.9
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.13
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.10
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.5.8
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
ruamel-yaml==0.15.46
rviz==1.12.16
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
sensor-msgs==1.12.7
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
sound-play==0.3.3
sounddevice==0.3.13
SoundFile==0.10.2
soupsieve==1.8
SpeechRecognition==3.8.1
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
SQLAlchemy==1.3.1
statsmodels==0.9.0
sympy==1.3
tables==3.5.1
tblib==1.3.2
tensorboard==1.12.2
tensorflow-estimator==1.13.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.20
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
tf2-sensor-msgs==0.5.20
toolz==0.9.0
topic-tools==1.12.14
torch==1.1.0
torchvision==0.3.0
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
typed-ast==1.3.1
unicodecsv==0.14.1
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
word2number==1.1
wrapt==1.11.1
wurlitzer==1.0.2
xacro==1.11.3
xlrd==1.2.0
XlsxWriter==1.1.5
xlwt==1.3.0
zict==0.1.4
zipp==0.3.3
